Changed Base Class to TwoArgBasic<Boolean> and moved Relationals to logic.h and logic.cpp
