+++ symengine/CMakeLists.txt
@@ -39,7 +39,6 @@ set(SRC
     cwrapper.cpp
     printer.cpp
     real_double.cpp
-    relationals.cpp
     complex_double.cpp
     expand.cpp
     expression.cpp
@@ -142,7 +141,6 @@ set(HEADERS
     rational.h
     real_double.h
     real_mpfr.h
-    relationals.h
     rings.h
     series_flint.h
     series_generic.h
